The type of field numElements of struct IndexedFloat32Array to uint32.

+++ python/boomer/common/cpp/tuples.h
@@ -23,7 +23,7 @@ struct IndexedFloat32 {
  */
 struct IndexedFloat32Array {
     IndexedFloat32* data;
-    intp numElements;
+    uint32 numElements;
 };
 
 /**
